final criteria fix
