--- js/const.js
@@ -1,76 +1,0 @@
-const Coordinates = {
-  MAX_LATITUDE: 35.70000,
-  MIN_LATITUDE: 35.65000,
-  MAX_LONGITUDE: 139.80000,
-  MIN_LONGITUDE: 139.70000,
-};
-
-const TYPE = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel',
-];
-
-const CHECKIN = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const CHECKOUT = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-];
-
-const PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
-];
-
-const PHOTO_NUMBERS = [
-  '01',
-  '02',
-  '03',
-  '04',
-  '05',
-  '06',
-  '07',
-  '08',
-  '09',
-  '10',
-];
-
-const TITLES = [
-  'Шикарная вилла у порта',
-  'Апартаменты на берегу лазурного моря',
-  'Бунгало на отлельном острове',
-  'Отель в сердце города',
-  'Микро домик в горах',
-  'Дом в снежных горах',
-  'Квартира на окраине города',
-];
-
-const DESCRIPTIONS = [
-  'Вид с окна на высоте птичьего полета',
-  'Тихие соседи',
-  'Горка примяком из номера',
-  'Выход из виллы прямо на пляж',
-  'Домик в центре леса',
-  'Центр города, рукой подать до театра, кино и центра развлечений',
-  'Рядом дорожки для бега вдоль океана',
-];
-
-export {Coordinates, TYPE, CHECKIN, CHECKOUT, FEATURES, PHOTOS, PHOTO_NUMBERS, TITLES, DESCRIPTIONS};
